Wrap up the basic spritesheet/animation implementation.
Timing information still needs to be added to that animation frames for
this to really be usable. That entails an update to the
Animation.frameAt function for looking up which frame to display, but
otherwise this looks ready to be used. It is also in main.lua for the
time being, and will be moved out when working.